[WinCairo][WK2] fast/events/wheel/wheelevent-basic.html is timing out https://bugs.webkit.org/show_bug.cgi?id=238590
Reviewed by Don Olmstead.

This issue happened only if SPI_GETWHEELSCROLLLINES was less than
3. If I changed "Choose how many lines to scroll each time" to 1
or 2 in Windows Settings, this test passed.

* Shared/win/WebEventFactory.cpp:
(WebKit::WebEventFactory::createWebWheelEvent): Changed local
variables to float not to truncate the wheel delta.



Canonical link: https://commits.webkit.org/249075@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@292168 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit/Shared/win/WebEventFactory.cpp
@@ -413,10 +413,10 @@ WebWheelEvent WebEventFactory::createWebWheelEvent(HWND hWnd, UINT message, WPAR
 
     auto modifiers = modifiersForEvent(wParam);
 
-    int deltaX = 0;
-    int deltaY = 0;
-    int wheelTicksX = 0;
-    int wheelTicksY = 0;
+    float deltaX = 0;
+    float deltaY = 0;
+    float wheelTicksX = 0;
+    float wheelTicksY = 0;
 
     float delta = GET_WHEEL_DELTA_WPARAM(wParam) / static_cast<float>(WHEEL_DELTA);
     bool isMouseHWheel = (message == WM_MOUSEHWHEEL);
